Regular expression support handling via UTF-8 in the locale
